fix(clippy): ratchet slice 1/5 — apr-cli, aprender-core, aprender-serve residual (Refs #4153)
Fixes the findings in code, never with #[allow] (0 added). The member
baseline loses 65 rows and 959 findings and gains none; the gate reads
"within baseline [workspace] (37 keys, 2318 findings)".

The guard's self-test mutated crates/aprender-build-sha, which exists only
on batch/0.70.0, so on this base it died with rc=101 before planting anything.
It now plants in aprender-common, which is on main:
"SELF-TEST PASSED: unchanged tree -> green; planted needless_return -> RED".

+++ crates/aprender-core/src/format/converter/tests/dogfood_2392.rs
@@ -265,7 +265,7 @@ mod finding_3_q4k_plan_is_shape_aware {
                 .tensor_names()
                 .iter()
                 .filter_map(|n| reader.get_tensor(n))
-                .map(|e| e.size as u64)
+                .map(|e| e.size)
                 .sum()
         };
         let _ = std::fs::remove_dir_all(&dir);
